+++ src/include/EntityManager.h
@@ -21,7 +21,7 @@ public:
     EntityManager();
     void update();
     void sort();
-    std::shared_ptr<Entity> addEntity(const std::string & tag, const size_t &layer);
+    std::shared_ptr<Entity> addEntity(const std::string & tag, const Uint8 &layer);
     void removeDeadEntities(EntityVec & vec);
     EntityVec getEntities();
     EntityVec getEntities(std::string tag);
